opencv-art/requirements.txt
numpy
opencv-python
opencv-contrib-python
progressbar2
scikit-learn
scipy

